PreferencesUserSettingsTab: Move the event index UI into a separate component.
